More changes to MDT CSV dumper for ACTS standalone interoperability

See merge request atlas/athena!67598

+++ MuonSpectrometer/MuonPhaseII/MuonCnv/MuonHitCsvDump/src/CsvMuonSimHitDumperMuonCnv.h
@@ -38,5 +38,5 @@
    const MuonGMR4::MuonDetectorManager* m_r4DetMgr{nullptr};
   PublicToolHandle<MuonGMR4::IMuonStationLayerSurfaceTool> m_surfaceProvTool{this, "LayerGeoTool", ""};
 
-  int m_event = 0;
+  size_t m_event = 0;
 };
